+++ src/ska_dlm_client/register_storage_location/main.py
@@ -26,7 +26,7 @@ LOCATION_COUNTRY = os.getenv("LOCATION_COUNTRY", LocationCountry.AU.value)
 LOCATION_CITY = os.getenv("LOCATION_CITY", "Perth")
 LOCATION_FACILITY = os.getenv("LOCATION_FACILITY", "local")
 TARGET_ROOT = os.getenv("TARGET_ROOT", "/dlm-archive")
-TARGET_PHASE = os.getenv("TARGET_PHASE", "SOLID")
+TGT_STORAGE_PHASE = os.getenv("TARGET_PHASE", "SOLID")  # "target phase" is ambigious.
 RCLONE_CONFIG_TARGET = {
     "name": "dlm-archive",
     "type": "alias",
@@ -231,7 +231,7 @@ def setup_testing(api_configuration: Configuration):
     storage_id = get_or_init_storage(
         storage_name=RCLONE_CONFIG_TARGET["name"],
         storage_url=storage_url,
-        storage_phase=TARGET_PHASE,
+        storage_phase=TGT_STORAGE_PHASE,
         api_configuration=api_configuration,
         storage_root_directory=TARGET_ROOT,
         the_location_id=location_id,